MAKE pdfconduit.internals  - add base.py with base pdfconduit functionality
